+++ packages/app/src/cli/services/generate/extension.ts
@@ -12,7 +12,11 @@ import {
 import {AppInterface} from '../../models/app/app.js'
 import {mapExtensionTypeToExternalExtensionType} from '../../utilities/extensions/name-mapper.js'
 import {error, file, git, path, string, template, ui, environment} from '@shopify/cli-kit'
-import {addNPMDependenciesIfNeeded, DependencyVersion} from '@shopify/cli-kit/node/node-package-manager'
+import {
+  addNPMDependenciesIfNeeded,
+  addResolutionOrOverride,
+  DependencyVersion,
+} from '@shopify/cli-kit/node/node-package-manager'
 import {fileURLToPath} from 'url'
 import stream from 'node:stream'
 
@@ -80,6 +84,7 @@ async function uiExtensionInit({
         title: 'Install additional dependencies',
         task: async (_, task) => {
           task.title = 'Installing additional dependencies...'
+          await addResolutionOrOverrideIfNeeded(app.directory, extensionFlavor)
           const requiredDependencies = getRuntimeDependencies({extensionType, extensionFlavor})
           await addNPMDependenciesIfNeeded(requiredDependencies, {
             packageManager: app.packageManager,
@@ -127,6 +132,7 @@ async function uiExtensionInit({
 
           if (extensionFlavor) {
             await changeIndexFileExtension(extensionDirectory, extensionFlavor)
+            await removeUnwantedTemplateFilesPerFlavor(extensionDirectory, extensionFlavor)
           }
 
           task.title = `${getExtensionOutputConfig(extensionType).humanKey} extension generated`
@@ -177,6 +183,14 @@ async function changeIndexFileExtension(extensionDirectory: string, extensionFla
       path.join(extensionDirectory, 'src/index'),
       path.join(extensionDirectory, `src/index.${fileExtension}`),
     )
+  }
+}
+
+async function removeUnwantedTemplateFilesPerFlavor(extensionDirectory: string, extensionFlavor: ExtensionFlavor) {
+  // tsconfig.json file is only needed in extension folder to inform the IDE
+  // About the `react-jsx` tsconfig option, so IDE don't complain about missing react import
+  if (extensionFlavor !== 'typescript-react') {
+    await file.remove(path.join(extensionDirectory, 'tsconfig.json'))
   }
 }
 
@@ -236,6 +250,12 @@ async function ensureExtensionDirectoryExists({name, app}: {name: string; app: A
   }
   await file.mkdir(extensionDirectory)
   return extensionDirectory
+}
+
+async function addResolutionOrOverrideIfNeeded(directory: string, extensionFlavor?: ExtensionFlavor) {
+  if (extensionFlavor === 'typescript-react') {
+    await addResolutionOrOverride(directory, {'@types/react': versions.reactTypes})
+  }
 }
 
 export default extensionInit