+++ packages/cli-hydrogen/src/cli/services/deploy.test.ts
@@ -15,6 +15,7 @@ const deployConfig: DeployConfig = {
 
 const reqDeployConfig: ReqDeployConfig = {
   ...deployConfig,
+  pathToBuild: '',
   commitMessage: 'commitMessage',
   commitAuthor: 'commitAuthor',
   commitSha: 'commitSha',
@@ -77,20 +78,87 @@ describe('deployToOxygen()', () => {
     expect(mockedHealthCheck).toHaveBeenCalledWith('previewURL')
   })
 
-  it('retries on error for healthCheck, uploadDeployment and createDeployment', async () => {
-    mockedFillDeployConfig.mockResolvedValue(reqDeployConfig)
-    mockedCreateDeployment
-      .mockRejectedValueOnce(new Error())
-      .mockResolvedValue({deploymentID: 'deploymentID', assetBaseURL: 'assetBaseURL'})
-    mockedBuildTaskList.mockReturnValue([])
-    mockedUploadDeployment.mockRejectedValueOnce(new Error()).mockResolvedValue('previewURL')
-    mockedHealthCheck.mockRejectedValueOnce(new Error()).mockRejectedValueOnce(new Error())
+  describe('when error is recoverable', async () => {
+    it('retries on error for healthCheck, uploadDeployment and createDeployment', async () => {
+      mockedFillDeployConfig.mockResolvedValue(reqDeployConfig)
+      mockedCreateDeployment
+        .mockRejectedValueOnce(new Error())
+        .mockResolvedValue({deploymentID: 'deploymentID', assetBaseURL: 'assetBaseURL'})
+      mockedBuildTaskList.mockReturnValue([])
+      mockedUploadDeployment.mockRejectedValueOnce(new Error()).mockResolvedValue('previewURL')
+      mockedHealthCheck.mockRejectedValueOnce(new Error()).mockRejectedValueOnce(new Error())
 
-    await deployToOxygen(deployConfig)
+      await deployToOxygen(deployConfig)
 
-    expect(mockedCreateDeployment).toHaveBeenCalledTimes(2)
-    expect(mockedUploadDeployment).toHaveBeenCalledTimes(2)
-    expect(mockedHealthCheck).toHaveBeenCalledTimes(3)
+      expect(mockedCreateDeployment).toHaveBeenCalledTimes(2)
+      expect(mockedUploadDeployment).toHaveBeenCalledTimes(2)
+      expect(mockedHealthCheck).toHaveBeenCalledTimes(3)
+    })
+
+    it('does retry for createDeployment', async () => {
+      mockedFillDeployConfig.mockResolvedValue(reqDeployConfig)
+      mockedCreateDeployment
+        .mockRejectedValueOnce(new Error('recoverable'))
+        .mockRejectedValueOnce(new Error('also recoverable'))
+        .mockResolvedValue({deploymentID: 'deploymentID', assetBaseURL: 'assetBaseURL'})
+
+      await deployToOxygen(deployConfig)
+
+      expect(mockedCreateDeployment).toHaveBeenCalledTimes(3)
+    })
+
+    it('does retry for uploadDeployment', async () => {
+      mockedFillDeployConfig.mockResolvedValue(reqDeployConfig)
+      mockedCreateDeployment.mockResolvedValue({deploymentID: 'deploymentID', assetBaseURL: 'assetBaseURL'})
+      mockedBuildTaskList.mockReturnValue([])
+      mockedUploadDeployment
+        .mockRejectedValueOnce(new Error('recoverable'))
+        .mockRejectedValueOnce(new Error('also recoverable'))
+        .mockResolvedValue('previewURL')
+
+      await deployToOxygen(deployConfig)
+
+      expect(mockedCreateDeployment).toHaveBeenCalledTimes(1)
+      expect(mockedUploadDeployment).toHaveBeenCalledTimes(3)
+    })
+  })
+
+  describe('when error is unrecoverable', async () => {
+    it('throws an exception on unrecoverable createDeployment error', async () => {
+      mockedFillDeployConfig.mockResolvedValue(reqDeployConfig)
+      mockedCreateDeployment.mockRejectedValueOnce(new Error('Unrecoverable: Should not retry'))
+
+      await expect(() => {
+        return deployToOxygen(deployConfig)
+      }).rejects.toThrowError('Could not create deployment on Oxygen. Unrecoverable: Should not retry')
+      expect(mockedCreateDeployment).toHaveBeenCalledTimes(1)
+    })
+
+    it('throws an exception on unrecoverable uploadDeployment error', async () => {
+      mockedFillDeployConfig.mockResolvedValue(reqDeployConfig)
+      mockedCreateDeployment.mockResolvedValue({deploymentID: 'deploymentID', assetBaseURL: 'assetBaseURL'})
+      mockedBuildTaskList.mockReturnValue([])
+      mockedUploadDeployment.mockRejectedValue(new Error('Unrecoverable: Should not retry'))
+
+      await expect(() => {
+        return deployToOxygen(deployConfig)
+      }).rejects.toThrowError('Uploading files to Oxygen failed. Unrecoverable: Should not retry')
+
+      expect(mockedCreateDeployment).toHaveBeenCalledTimes(1)
+      expect(mockedUploadDeployment).toHaveBeenCalledTimes(1)
+    })
+  })
+
+  describe('when the deploy config has pathToBuild set', () => {
+    it('skips build step', async () => {
+      mockedFillDeployConfig.mockResolvedValue({...reqDeployConfig, pathToBuild: '/some/test/path'})
+      mockedCreateDeployment.mockResolvedValue({deploymentID: 'deploymentID', assetBaseURL: 'assetBaseURL'})
+      mockedBuildTaskList.mockReturnValue([])
+
+      await deployToOxygen(deployConfig)
+
+      expect(mockedBuildTaskList).not.toHaveBeenCalled()
+    })
   })
 
   describe('when the deploy config has healthCheck set to false', () => {