+++ packages/cli-kit/src/id.ts
@@ -2,7 +2,7 @@ import {randomUUID} from 'crypto'
 
 /**
  * Generates and returns a random UUID.
- * @returns {string} The random UUID generated.
+ * @returns The random UUID generated.
  */
 export const generateRandomUUID = (): string => {
   return randomUUID()